requirements.txt
google-generativeai>=0.4.0
PyYAML>=6.0
python-dotenv>=1.0.0
